Rewrite Command Registration
This cleans the World init function up, Do add every Command and its logic there was a messy way, Now its clean
