+++ .gitignore
@@ -2,6 +2,8 @@
 /log
 /tmp
 
+/.vscode
+
 /.idea
 
 *.log
