+++ bb_server/src/world/init.rs
@@ -1,6 +1,8 @@
 use super::{World, WorldManager};
 use crate::{
-  command::{Arg, Command, Parser, StringType},
+  command::{
+    custom::add_custom_commands, vanilla::add_vanilla_commands, Arg, Command, Parser, StringType,
+  },
   entity,
   player::Player,
 };
@@ -18,164 +20,26 @@ impl World {
       self.load_from_disk(&std::path::PathBuf::new().join(&self.config().vanilla.path)).unwrap();
     }
 
-    let mut c = Command::new("say");
-    c.add_arg("text", Parser::String(StringType::Greedy));
-    self.commands().add(c, |world, _, args| {
-      world.broadcast(format!("[Server] {}", args[1].str()).as_str());
-    });
+    // TODO: Make this a way to toggle this on/off
+    add_vanilla_commands(self.commands());
+    add_custom_commands(self.commands());
 
-    let mut c = Command::new("fill");
-    c.add_lit("rect")
-      .add_arg("min", Parser::BlockPos)
-      .add_arg("max", Parser::BlockPos)
-      .add_arg("block", Parser::BlockState);
-    c.add_lit("circle")
-      .add_arg("center", Parser::BlockPos)
-      .add_arg("radius", Parser::Float { min: Some(0.0), max: None })
-      .add_arg("block", Parser::BlockState);
-    c.add_lit("sphere")
-      .add_arg("center", Parser::BlockPos)
-      .add_arg("radius", Parser::Float { min: Some(0.0), max: None })
-      .add_arg("block", Parser::BlockState);
-    self.commands().add(c, |world, _, args| {
-      // args[0] is `fill`
-      match args[1].lit() {
-        "rect" => {
-          let min = args[2].pos();
-          let max = args[3].pos();
-          let block = args[4].block();
-          let (min, max) = min.min_max(max);
-          let w = world.default_world();
-          w.fill_rect_kind(min, max, block).unwrap();
-        }
-        "circle" => {
-          let pos = args[2].pos();
-          let radius = args[3].float();
-          let block = args[4].block();
-          let w = world.default_world();
-          w.fill_circle_kind(pos, radius, block).unwrap();
-        }
-        "sphere" => {
-          let pos = args[2].pos();
-          let radius = args[3].float();
-          let block = args[4].block();
-          let w = world.default_world();
-          w.fill_sphere_kind(pos, radius, block).unwrap();
-        }
-        _ => unreachable!(),
-      }
-    });
-    fn handle_gamemode(wm: &Arc<WorldManager>, runner: Option<&Arc<Player>>, args: Vec<Arg>) {
-      let gm = match &args[1] {
-        Arg::Literal(lit) => match lit.as_str() {
-          "survival" | "s" => GameMode::Survival,
-          "creative" | "c" => GameMode::Creative,
-          "adventure" | "a" => GameMode::Adventure,
-          "spectator" | "sp" => GameMode::Spectator,
-          _ => unreachable!(),
-        },
-        Arg::Int(num) => GameMode::from_id(*num as u8),
-        _ => unreachable!(),
-      };
-      if let Some(arg) = args.get(2) {
-        for world in wm.worlds().iter() {
-          for target in arg.entity().iter(&world.entities(), runner) {
-            if let Some(p) = target.as_player() {
-              p.set_game_mode(gm)
-            }
-          }
-        }
-      } else if let Some(player) = runner {
-        player.set_game_mode(gm);
-      } else {
-        // TODO: Send error saying they need to specify a target
-      }
-    }
-    for name in ["gamemode", "gm"] {
-      let mut c = Command::new(name);
-      c.add_lit("survival")
-        .add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("creative")
-        .add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("adventure")
-        .add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("spectator")
-        .add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("s").add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("c").add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("a").add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_lit("sp").add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      c.add_arg("mode", Parser::Int { min: Some(0), max: Some(3) });
-      self.commands().add(c, handle_gamemode);
-    }
+    // we should move this Code to the Proxy
 
-    let add_specific_game_mode = |name: &'static str, gm: GameMode| {
-      let mut c = Command::new(name);
-      c.add_arg_opt("target", Parser::Entity { single: false, only_players: true });
-      self.commands().add(c, move |wm, runner, args| {
-        if let Some(arg) = args.get(1) {
-          for world in wm.worlds().iter() {
-            for target in arg.entity().iter(&world.entities(), runner) {
-              if let Some(p) = target.as_player() {
-                p.set_game_mode(gm)
-              }
-            }
-          }
-        } else if let Some(player) = runner {
-          player.set_game_mode(gm);
-        } else {
-          // TODO: Send error saying they need to specify a target
-        }
-      });
-    };
-
-    add_specific_game_mode("gms", GameMode::Survival);
-    add_specific_game_mode("gmc", GameMode::Creative);
-    add_specific_game_mode("gma", GameMode::Adventure);
-    add_specific_game_mode("gmsp", GameMode::Spectator);
-
-    let c = Command::new("fly");
-    self.commands().add(c, |_, player, _| {
-      if let Some(p) = player {
-        p.set_flying_allowed(!p.flying_allowed());
-      }
-    });
-    let mut c = Command::new("flyspeed");
-    c.add_arg("multiplier", Parser::Float { min: None, max: None });
-    self.commands().add(c, |_, player, args| {
-      // args[0] is `flyspeed`
-      let v = args[1].float();
-      if let Some(p) = player {
-        p.set_flyspeed(v);
-      }
-    });
-
-    let mut c = Command::new("summon");
-    c.add_arg("entity", Parser::EntitySummon);
-    self.commands().add(c, |_, player, args| {
-      // args[0] is `summon`
-      let ty = args[1].entity_summon();
-      if let Some(p) = player {
-        let eid = p.world().summon(ty, p.pos());
-        info!("eid of mob: {}", eid);
-        p.send_message(Chat::new(format!("summoned {ty:?}")));
-      }
-    });
-
-    let mut c = Command::new("join");
-    c.add_arg("addr", Parser::String(StringType::Word));
-    self.commands().add(c, |_, player, args| {
-      if let Some(p) = player {
-        let addr = match args[1].str().parse::<std::net::SocketAddr>() {
-          Ok(a) => a,
-          Err(e) => {
-            p.send_message(Chat::new(e.to_string()));
-            return;
-          }
-        };
-        p.switch_to(SwitchMode::Loading, vec![addr]);
-      }
-    });
+    // let mut c = Command::new("join");
+    // c.add_arg("addr", Parser::String(StringType::Word));
+    // self.commands().add(c, |_, player, args| {
+    //   if let Some(p) = player {
+    //     let addr = match args[1].str().parse::<std::net::SocketAddr>() {
+    //       Ok(a) => a,
+    //       Err(e) => {
+    //         p.send_message(Chat::new(e.to_string()));
+    //         return;
+    //       }
+    //     };
+    //     p.switch_to(SwitchMode::Loading, vec![addr]);
+    //   }
+    // });
 
     info!("generating terrain...");
     /*
